solve dp 1 chang package name
