+++ baekjoon_type/dp/b10870.java
@@ -1,26 +1,24 @@
-package dynamic;
+package dp;
 
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
 
-public class b9095 {
+public class b10870 {
 	public static void main(String[] args) throws NumberFormatException, IOException {
 		BufferedReader br =new BufferedReader(new InputStreamReader(System.in));
 		int n = Integer.parseInt(br.readLine());
+		System.out.println(fibo(n));
 		
-		int [] dp = new int[11];
-		dp[0] = 0;
-		dp[1] = 1;
-		dp[2] = 2;
-		dp[3] = 4;
-		for(int j = 4; j <=10; j++) {
-			dp[j] = dp[j-1] + dp[j-2] + dp[j-3];
+	}
+
+	private static int fibo(int n) {
+		if(n == 0) {
+			return 0;
 		}
-		
-		for(int i = 0; i < n; i++) {
-			int temp = Integer.parseInt(br.readLine());
-			System.out.println(dp[temp]);
+		if(n == 1) {
+			return 1;
 		}
+		return fibo(n-1) + fibo(n-2);
 	}
 }
