+++ baekjoon_type/dp/b11053.java
@@ -1,36 +1,40 @@
-package dynamic;
+package dp;
 
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.util.StringTokenizer;
 
-public class b1912 {
-	
+public class b11053 {
 	public static void main(String[] args) throws NumberFormatException, IOException {
 		BufferedReader br =new BufferedReader(new InputStreamReader(System.in));
 		int n = Integer.parseInt(br.readLine());
-		
 		int arr[] = new int[n];
 		int dp[] = new int[n];
 		StringTokenizer st = new StringTokenizer(br.readLine());
-		for(int i=0; i< n; i++) {
+		for(int i = 0; i  < n ; i ++) {
 			arr[i] = Integer.parseInt(st.nextToken());
 		}
-		dp[0] = arr[0];
-		for(int i = 1; i<n; i++) {
-			dp[i] = Math.max(dp[i-1]+arr[i], arr[i]);
-		}
 		
-		int max = -1000;
-		for(int i = 0; i < n; i++) {
-			if(max < dp[i]) {
+		for(int i =0; i < n; i++) {
+			dp[i] = 1;
+			for(int j =0; j < i; j++) {
+				if(arr[i] > arr[j] && dp[i] < dp[j]+1) {
+					dp[i] = dp[j]+1;
+				}
+			}
+		}
+		int max = 0;
+		for(int i = 0; i < n; i ++) {
+			if(dp[i] > max) {
 				max = dp[i];
 			}
 		}
 		System.out.println(max);
+		
 	}
 
 	
 }
+
