+++ baekjoon_type/dp/b11726.java
@@ -1,24 +1,20 @@
-package dynamic;
+package dp;
 
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
 
-public class b10870 {
+public class b11726 {
+	//down-top
 	public static void main(String[] args) throws NumberFormatException, IOException {
 		BufferedReader br =new BufferedReader(new InputStreamReader(System.in));
 		int n = Integer.parseInt(br.readLine());
-		System.out.println(fibo(n));
-		
-	}
-
-	private static int fibo(int n) {
-		if(n == 0) {
-			return 0;
+		int [] dp = new int[n+2];
+		dp[1] = 1;
+		dp[2] = 2;
+		for(int i =3; i <=n; i++) {
+			dp[i] = (dp[i-1]+dp[i-2]) %10007;
 		}
-		if(n == 1) {
-			return 1;
-		}
-		return fibo(n-1) + fibo(n-2);
+		System.out.println(dp[n]);
 	}
 }
