+++ baekjoon_type/dp/b1699.java
@@ -1,4 +1,4 @@
-package dynamic;
+package dp;
 
 
 import java.io.BufferedReader;
@@ -7,31 +7,38 @@ import java.io.InputStreamReader;
 import java.util.StringTokenizer;
 
 
-public class b11052 {
+public class b1699 {
 
 	static int[] dp;
 	static int [] arr;
-	static int n;
-	static int west;
-	static int east;
+	static int n, k;
 	public static void main(String[] args) throws NumberFormatException, IOException {
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
-		n = Integer.parseInt(br.readLine());
-			
-		arr = new int[n+1];
-		dp = new int[n+1];
+		
 		StringTokenizer st = new StringTokenizer(br.readLine());
 		
-		for(int i = 1; i  <= n; i++) {
-			arr[i] = Integer.parseInt(st.nextToken());
-		}
+		n = Integer.parseInt(st.nextToken());
 
-		for(int i = 1; i <= n ; i++) {
-			for(int j = 1; j <= i; j++) {
-				dp[i] = Math.max(dp[i], arr[j] + dp[i-j]);
-			}
+		dp = new int[n+1];
+		
+		for(int i =1; i <=n;i++) {
+			dp[i] = i;
 		}
+		
+		for(int i = 1; i <=n; i++) {
+			
+			for(int j =1; j*j <=i; j++) {
+				if(dp[i] > dp[i-(j*j)]+1) {
+					dp[i] = dp[i-(j*j)]+1;
+				}
+			}
+			
+		}
+		
+		
 		System.out.println(dp[n]);
+		
+		
 	}
 }
