+++ baekjoon_type/dp/b1904.java
@@ -1,4 +1,4 @@
-package dynamic;
+package dp;
 
 
 import java.io.BufferedReader;
