+++ baekjoon_type/dp/b1912.java
@@ -1,0 +1,36 @@
+package dp;
+
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStreamReader;
+import java.util.StringTokenizer;
+
+public class b1912 {
+	
+	public static void main(String[] args) throws NumberFormatException, IOException {
+		BufferedReader br =new BufferedReader(new InputStreamReader(System.in));
+		int n = Integer.parseInt(br.readLine());
+		
+		int arr[] = new int[n];
+		int dp[] = new int[n];
+		StringTokenizer st = new StringTokenizer(br.readLine());
+		for(int i=0; i< n; i++) {
+			arr[i] = Integer.parseInt(st.nextToken());
+		}
+		dp[0] = arr[0];
+		for(int i = 1; i<n; i++) {
+			dp[i] = Math.max(dp[i-1]+arr[i], arr[i]);
+		}
+		
+		int max = -1000;
+		for(int i = 0; i < n; i++) {
+			if(max < dp[i]) {
+				max = dp[i];
+			}
+		}
+		System.out.println(max);
+	}
+
+	
+}
+
