+++ baekjoon_type/dp/b2748.java
@@ -1,31 +1,22 @@
-package dynamic;
+package dp;
 
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
 
-public class b2579 {
+public class b2748 {
 	
 	public static void main(String[] args) throws NumberFormatException, IOException {
 		BufferedReader br =new BufferedReader(new InputStreamReader(System.in));
 		int n = Integer.parseInt(br.readLine());
-		int [] arr = new int[n+1];
-		int [] dp = new int[n+1];
-		for(int i = 1; i <= n; i++) {
-			arr[i] = Integer.parseInt(br.readLine());
-		}
-		dp[1] = arr[1];
+		long[]dp = new long[n+1];
+		dp[0] = 0;
+		dp[1] = 1;
 		if(n >=2) {
-			dp[2] = arr[1] + arr[2];
-		}
-		
-		for(int i = 3; i <=n; i++) {
-			dp[i] = Math.max(dp[i-2], dp[i-3] + arr[i-1])  + arr[i];
+			for(int i = 2; i<=n;i++) {
+				dp[i] = dp[i-1] +dp[i-2];
+			}
 		}
 		System.out.println(dp[n]);
-		
 	}
-
-	
 }
-
