+++ baekjoon_type/dp/b9184.java
@@ -1,4 +1,4 @@
-package dynamic;
+package dp;
 
 import java.io.BufferedReader;
 import java.io.IOException;
