+++ baekjoon_type/dp/b9461.java
@@ -1,4 +1,4 @@
-package dynamic;
+package dp;
 
 
 import java.io.BufferedReader;
@@ -6,7 +6,7 @@ import java.io.IOException;
 import java.io.InputStreamReader;
 
 
-public class b2193 {
+public class b9461 {
 
 	static long[] dp;
 	static int arr[];
@@ -16,23 +16,23 @@ public class b2193 {
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 		
 		int n = Integer.parseInt(br.readLine());
-		dp  = new long[n+1];
-		
-		dp[0] =0;
-		dp[1] = 1;
-		if(n>=2) {
-			dp[2]  =1;
+	
+		while(n--> 0) {
+			int a= Integer.parseInt(br.readLine());
+			dp = new long[a+1];
+			dp[0] = 0; 
+			dp[1] = 1;
+			if(a>=2) {dp[2] = 1;};
+			for(int i = 3; i < a+1; i++) {
+				dp[i] = dp[i-2] + dp[i-3];
+			}
+			System.out.println(dp[a]);
 		}
-		
-		for(int i = 3; i < n+1; i ++) {
-			dp[i] = dp[i-2] + dp[i-1];
-		}
-		System.out.println(dp[n]);
-		
 		
 	}
 }
 
 	
 	
+
